Refactor package naming from 'actividadOcupacion' to 'ocupacion' across documentation and code references. Updated HTML files, index files, and package summaries to reflect the new package structure, ensuring consistency in links and descriptions.

+++ src/main/java/com/coagronet/evaluacion/repositories/EvaluacionRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.coagronet.evaluacion.Evaluacion;
 
 @Repository
-public interface EvaluacionRepository extends JpaRepository<Evaluacion, Integer> {
+public interface EvaluacionRepository extends JpaRepository<Evaluacion, Long> {
 
     List<Evaluacion> findByTipoEvaluacionId(
             Integer tipoEvaluacionId);
